+++ frontend/src/app/website.service.ts
@@ -1,13 +1,15 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable, catchError, of, tap } from 'rxjs';
 import { Page, Website } from './website';
+import { EXAMPLE_SITES } from './MOCKSITES';
+
 
 @Injectable({
   providedIn: 'root'
 })
 export class WebsiteService {
-
+  
   private websiteURL = 'api/website';
   private websitesURL = 'api/websites';
 
@@ -19,11 +21,20 @@ export class WebsiteService {
 
   getWebsiteById(id: string): Observable<Website> {
     const url = `${this.websiteURL}/${id}`;
-    return this.http.get<Website>(url);
+    return this.http.get<Website>(url)
+          .pipe(
+            catchError(this.handleError<Website>('getWebsite'))
+            );
   }
 
   getWebsites(): Observable<Website[]> {
-    return this.http.get<Website[]>(this.websitesURL);
+    // hopefully we wont need these
+    // const sites = of(EXAMPLE_SITES);
+    // return sites;
+    return this.http.get<Website[]>(this.websiteURL)
+          .pipe(
+            catchError(this.handleError<Website[]>('getWebsites'))
+            );
   }
 
   addWebsite(website: Website): Observable<Website> {
@@ -34,4 +45,21 @@ export class WebsiteService {
     const url = `${this.websiteURL}/${websiteId}`;
     return this.http.post<Page>(url, page, this.httpOptions);
   }
+
+  /* dw about this until sprint 2 */
+  deleteWebsite(){}
+
+  deletePageToWebsite(){}
+
+
+  /** error handling */
+  private handleError<T>(operation = 'operation', result?: T) {
+		return (error: any): Observable<T> => {
+
+		// TODO: send the error to remote logging infrastructure
+		console.error(error); // log to console instead
+		// Let the app keep running by returning an empty result.
+		return of(result as T);
+		};
+	}
 }
